requirements.txt
streamlit==1.32.2
opencv-python==4.8.0.76  # Using non-headless version for better compatibility
tensorflow-cpu==2.15.0
numpy==1.26.4
Pillow==10.2.0
absl-py==2.1.0
protobuf==3.20.3
pyyaml==6.0.1